+++ catkin_ws/src/gait_generation/step_test/package.xml
@@ -45,11 +45,13 @@
   <build_depend>rospy</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>tf</build_depend>
+  <build_depend>trajectory_planner</build_depend>
   <run_depend>control_msgs</run_depend>
   <run_depend>roscpp</run_depend>
   <run_depend>rospy</run_depend>
   <run_depend>std_msgs</run_depend>
   <run_depend>tf</run_depend>
+  <run_depend>trajectory_planner</run_depend>
 
 
   <!-- The export tag contains other, unspecified, tags -->
