+++ catkin_ws/src/hardware/nimbro_op_model/robots/simple_nimbro_op_hull_fixed.xml
@@ -2,7 +2,7 @@
   <material name="black">
     <color rgba="0.1 0.1 0.1 1" />
   </material>
-  <!-- <link name="world"/> -->
+  <link name="world"/>
   <link name="trunk_link">
     <inertial>
       <origin xyz="-0.016856 -0.00019822 0.019225" rpy="0 0 0" />
@@ -24,11 +24,11 @@
       </geometry>
     </collision>
   </link>
-  <!-- <joint name="fixed_world" type="fixed">
+  <joint name="fixed_world" type="fixed">
     <origin xyz="0 0 1" rpy="0 0 0" />
     <parent link="world" />
     <child link="trunk_link" />
-  </joint> -->
+  </joint>
   <link name="right_hip_yaw_link">
     <inertial>
       <origin xyz="4.7878E-16 2.8449E-16 -0.013803" rpy="0 0 0" />
