+++ catkin_ws/src/surge_et_ambula/launch/humanoid_hardware_simul.launch
@@ -1,5 +1,7 @@
 <launch>
-    <param name="robot_description" command="cat $(find nimbro_op_model)/robots/simple_nimbro_op_hull.xml" />
+    <arg name="fixed" default="false"/>
+    <param name="robot_description" command="cat $(find nimbro_op_model)/robots/simple_nimbro_op_hull.xml" unless="$(arg fixed)"/>
+    <param name="robot_description" command="cat $(find nimbro_op_model)/robots/simple_nimbro_op_hull_fixed.xml" if="$(arg fixed)"/>
     <param name="robot_description_semantic" textfile="$(find nimbro_op_srdf)/config/nimbro_op_hull.srdf" />
     <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher"/>
